requirements.txt
python-dotenv
spicy
pandas
rasterio
geopandas
shapely
fiona
matplotlib
numpy
scikit-learn
pyarrow